[YS-224] refactor: AreaTooltip 스타일 코드 분리
