--- src/app/join/components/Participant/JoinInfoStep/AreaTooltip.tsx
@@ -1,61 +1,0 @@
-import { css, keyframes, Theme } from '@emotion/react';
-import * as Tooltip from '@radix-ui/react-tooltip';
-
-import Icon from '@/components/Icon';
-
-const AreaTooltip = () => {
-  return (
-    <Tooltip.Provider>
-      <Tooltip.Root>
-        <Tooltip.Trigger asChild>
-          <button>
-            <Icon icon="Information" width={18} height={18} />
-          </button>
-        </Tooltip.Trigger>
-        <Tooltip.Portal>
-          <Tooltip.Content side="bottom" sideOffset={2} css={tooltipContent}>
-            학교 소재지 등 자주 가는 지역을 추가로 입력할 수 있어요
-            <Tooltip.Arrow css={tooltipArrow} />
-          </Tooltip.Content>
-        </Tooltip.Portal>
-      </Tooltip.Root>
-    </Tooltip.Provider>
-  );
-};
-
-export default AreaTooltip;
-
-const tooltipContent = (theme: Theme) => css`
-  ${theme.fonts.label.medium.M13};
-  width: 20rem;
-  left: 2.4rem;
-  background-color: ${theme.colors.field01};
-  border-radius: 0.6rem;
-  padding: 0.8rem 1.6rem;
-  color: ${theme.colors.text05};
-  box-shadow: 0 4px 8px rgba(16, 17, 18, 0.1);
-  border: 0.15rem solid ${theme.colors.line01};
-
-  user-select: none;
-  animation-duration: 100ms;
-  animation-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
-
-  &[data-state='delayed-open'][data-side='bottom'] {
-    animation-name: ${slideDownAndFade};
-  }
-`;
-
-const tooltipArrow = (theme: Theme) => css`
-  fill: ${theme.colors.field01};
-`;
-
-const slideDownAndFade = keyframes`
-	from {
-		opacity: 0;
-		transform: translateY(-2px);
-	}
-	to {
-		opacity: 1;
-		transform: translateY(0);
-	}
-`;
