+++ src/app/join/components/Participant/JoinInfoStep/AreaTooltip/AreaTooltip.tsx
@@ -1,0 +1,26 @@
+import * as Tooltip from '@radix-ui/react-tooltip';
+
+import Icon from '@/components/Icon';
+import { tooltipArrow, tooltipContent } from './AreaTooltip.styles';
+
+const AreaTooltip = () => {
+  return (
+    <Tooltip.Provider>
+      <Tooltip.Root>
+        <Tooltip.Trigger asChild>
+          <button>
+            <Icon icon="Information" width={18} height={18} />
+          </button>
+        </Tooltip.Trigger>
+        <Tooltip.Portal>
+          <Tooltip.Content side="bottom" sideOffset={2} css={tooltipContent}>
+            학교 소재지 등 자주 가는 지역을 추가로 입력할 수 있어요
+            <Tooltip.Arrow css={tooltipArrow} />
+          </Tooltip.Content>
+        </Tooltip.Portal>
+      </Tooltip.Root>
+    </Tooltip.Provider>
+  );
+};
+
+export default AreaTooltip;
